dashboard/requirements_dynamo.txt
boto3==1.34.127
dash==2.17.1
dash_bootstrap_components==1.6.0
dynamodb-json==1.4.2
pandas==2.2.2
plotly==5.23.0
python-dotenv==1.0.1
pytz==2023.3.post1